detalles

+++ src/services/wiki/categories.ts
@@ -19,8 +19,6 @@ export async function updateTopicsCategories(ctx: AppContext) {
         categoryIds: getTopicCategories(t.props as unknown as TopicProp[])
     }))
 
-    console.log("topic categories", topicCategories)
-
     const t1 = Date.now();
 
     const allCategoryIds = [
@@ -32,13 +30,17 @@ export async function updateTopicsCategories(ctx: AppContext) {
         return;
     }
 
+
     await ctx.kysely.transaction().execute(async (trx) => {
-        console.log('Inserting categories.');
-        await ctx.kysely
-            .insertInto('TopicCategory')
-            .values(allCategoryIds.map((id) => ({id})))
-            .onConflict((oc) => oc.column('id').doNothing())
-            .execute();
+        try {
+            await trx
+                .insertInto('TopicCategory')
+                .values(allCategoryIds.map((id) => ({id})))
+                .onConflict((oc) => oc.column('id').doNothing())
+                .execute();
+        } catch (err) {
+            console.log(`Error inserting categories: ${err}.`);
+        }
 
         const topicCategoryValues = topicCategories.flatMap(({topicId, categoryIds}) =>
             categoryIds.map((categoryId) => ({
@@ -46,21 +48,15 @@ export async function updateTopicsCategories(ctx: AppContext) {
                 categoryId
             }))
         );
-        console.log("topic-category values",
-            topicCategoryValues.length,
-            topicCategoryValues.slice(0, 20)
-        )
 
         if (topicCategoryValues.length === 0) {
             console.log('No topic-category relationships to update.');
             return;
         }
 
-        console.log(topicCategoryValues.slice(0, 20))
-
         let existing: {topicId: string, categoryId: string}[] = []
         try {
-            existing = await ctx.kysely
+            existing = await trx
                 .selectFrom('TopicToCategory')
                 .select(['topicId', 'categoryId'])
                 .execute();
@@ -68,37 +64,54 @@ export async function updateTopicsCategories(ctx: AppContext) {
             console.log("Error getting existing relations", err)
             return
         }
+        const batchSize = 500;
 
         console.log("Inserting new relations")
-        await ctx.kysely
-            .insertInto('TopicToCategory')
-            .values(topicCategoryValues)
-            .onConflict((oc) => oc.columns(['topicId', 'categoryId']).doNothing())
-            .execute();
 
+        for(let i = 0; i < topicCategoryValues.length; i += batchSize) {
+            console.log(`Batch ${i}.`)
+            console.log(topicCategoryValues.slice(i, i + batchSize).slice(0, 5))
+            try {
+                await trx
+                    .insertInto('TopicToCategory')
+                    .values(topicCategoryValues.slice(i, i + batchSize))
+                    .onConflict((oc) => oc.columns(['topicId', 'categoryId']).doNothing())
+                    .execute();
+            } catch (err) {
+                console.log(`Error inserting relations: ${err}.`);
+            }
+        }
         const toDelete = existing.filter(r => !topicCategoryValues.some(v => v.topicId == r.topicId && v.categoryId == r.categoryId));
 
         console.log("Deleting old relations", toDelete.length)
         try {
-            const batch_size = 500;
-            for (let i = 0; i < toDelete.length; i += batch_size) {
-                console.log(`Deleting batch starting at ${i} of ${toDelete.length} relations.`)
-                const chunk = toDelete.slice(i, i + batch_size);
-                await ctx.kysely
-                    .deleteFrom('TopicToCategory')
-                    .where(({eb, refTuple, tuple}) =>
-                        eb(
-                            refTuple('topicId', 'categoryId'),
-                            'not in',
-                            chunk.map((r) => tuple(r.topicId, r.categoryId))
-                        )
-                    )
-                    .execute();
+            if(toDelete.length > 0) {
+                for (let i = 0; i < toDelete.length; i += batchSize) {
+                    console.log(`Deleting batch starting at ${i} of ${toDelete.length} relations.`)
+                    const chunk = toDelete.slice(i, i + batchSize);
+                    try {
+                        await trx
+                            .deleteFrom('TopicToCategory')
+                            .where(({eb, refTuple, tuple}) =>
+                                eb(
+                                    refTuple('topicId', 'categoryId'),
+                                    'in',
+                                    chunk.map((r) => tuple(r.topicId, r.categoryId))
+                                )
+                            )
+                            .execute();
+                    } catch (err) {
+                        console.log(`Error deleting batch: ${err}.`);
+                    }
+                }
             }
         } catch (err) {
             console.log("Error deleting old relations", err)
         }
     })
+
+
+
 
 
     console.log('Done after', Date.now() - t1);
